more fixes and comments from lawrence
