+++ src/include/duckdb/common/common.hpp
@@ -7,6 +7,5 @@
 //===----------------------------------------------------------------------===//
 
 #pragma once
-
 #include "duckdb/common/constants.hpp"
 #include "duckdb/common/helper.hpp"
