+++ src/include/duckdb/optimizer/estimated_properties.hpp
@@ -1,0 +1,32 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/optimizer/estimated_properties.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/common/unordered_map.hpp"
+#include "duckdb/common/unordered_set.hpp"
+#include "duckdb/optimizer/join_order/query_graph.hpp"
+#include "duckdb/optimizer/join_order/join_relation.hpp"
+#include "duckdb/parser/expression_map.hpp"
+#include "duckdb/planner/logical_operator_visitor.hpp"
+#include "duckdb/storage/statistics/distinct_statistics.hpp"
+#include "duckdb/planner/table_filter.hpp"
+#include "duckdb/catalog/catalog_entry/table_catalog_entry.hpp"
+
+namespace duckdb {
+
+class EstimatedProperties {
+public:
+	idx_t cost;
+	double cardinality;
+
+	unique_ptr<EstimatedProperties> Copy();
+
+private:
+};
+} // namespace duckdb
